+++ platformio.ini
@@ -18,6 +18,7 @@ lib_deps =
 	adafruit/Adafruit ADS1X15@^2.2.0
 	bblanchon/ArduinoJson@^6.18.3
 	ottowinter/ESPAsyncTCP-esphome@^1.2.3
-	arduino-libraries/Ethernet@^2.0.0
 monitor_port = COM6
 upload_port = 192.168.10.151
+src_filter = +<*>  -<Ethernet_Config.h> -<Ethernet_Config.cpp>
+
