+++ src/main.cpp
@@ -1,183 +1,37 @@
 #include <Arduino.h>
 #include <WiFi.h>
 #include <WebServer.h>
-
 #include <WiFiUdp.h>
 #include <ArduinoOTA.h>
 #include <Wire.h>
 #include <Adafruit_ADS1X15.h>
 #include <SPI.h>
 #include <ArduinoJson.h>
-#include <Ethernet.h>
+//#include <Ethernet.h>
 #include "Passwords.h"
-#include "Ethernet_Config.h"
+//#include "Ethernet_Config.h"
+//#include <EthernetWebServer.h>
+
+#define SDEF(sname, ...) S sname __VA_OPT__(= {__VA_ARGS__})
 
 WebServer server(80);
+// EthernetWebServer ethernetServer(80);
 
 void connectToStuff();
 void initOTA();
 void reconnect();
 
-WiFiClient espClient;
+// WiFiClient espClient;
 
 float voltageInputMultiplier[6][8];
 float voltageOutputMultiplier[4] = {1, 1, 1, 1};
 uint16_t voltageOutputOffset[4] = {2047, 2047, 2047, 2047};
 bool adsStatus[4], expanderStatus[8], voltageOutputsStatus[4];
 
-DynamicJsonDocument doc(256);
-
 #include "Analog_Inputs.h"
 #include "IO_expander.h"
 #include "Voltage_Outputs.h"
-
-void handleRoot()
-{
-    char temp[400];
-    int sec = millis() / 1000;
-    int min = sec / 60;
-    int hr = min / 60;
-
-    snprintf(temp, 400,
-
-             "<html>\
-  <head>\
-    <meta http-equiv='refresh' content='5'/>\
-    <title>ESP8266 Demo</title>\
-    <style>\
-      body { background-color: #cccccc; font-family: Arial, Helvetica, Sans-Serif; Color: #000088; }\
-    </style>\
-  </head>\
-  <body>\
-    <h1>Hello from ESP8266!</h1>\
-    <p>Uptime: %02d:%02d:%02d</p>\
-    <img src=\"/test.svg\" />\
-  </body>\
-</html>",
-
-             hr, min % 60, sec % 60);
-    server.sendHeader("Access-Control-Allow-Origin", "*");
-    server.send(200, "text/html", temp);
-}
-
-void handleAnalogInputs()
-{
-    if (server.method() != HTTP_POST)
-    {
-        digitalWrite(led, 1);
-        server.send(405, "text/plain", "Method Not Allowed");
-        digitalWrite(led, 0);
-    }
-    else
-    {
-        String temp;
-        String message = "{\"data\":[\n";
-        for (uint8_t i = 0; i < server.args(); i++)
-        {
-            if (server.argName(i) == "ADCREAD")
-            {
-                measureAndReturn(server.arg(i).toInt(), &temp);
-                message += "\t{ " + temp + " }";
-                if (i < server.args() - 1)
-                    message += ",\n";
-            }
-            if (server.argName(i) == "ADCRAW")
-            {
-                measureAndReturnRAW(server.arg(i).toInt(), &temp);
-                message += "\t{ " + temp + " }";
-                if (i < server.args() - 1)
-                    message += ",\n";
-            }
-        }
-        message += "\n]}";
-        server.sendHeader("Access-Control-Allow-Origin", "*");
-        server.send(200, "text/plain", message);
-    }
-}
-
-void handleAnalogOutputs()
-{
-    if (server.method() != HTTP_POST)
-    {
-        digitalWrite(led, 1);
-        server.send(405, "text/plain", "Method Not Allowed");
-        digitalWrite(led, 0);
-    }
-    else
-    {
-        String message = "{\"data\":[\n";
-        for (uint8_t i = 0; i < server.args(); i++)
-        {
-            if (server.argName(i).substring(0, 4) == "VSET")
-            {
-                setChannelVoltage((char)server.argName(i).substring(4).toInt(), server.arg(i).toFloat());
-            }
-        }
-        server.sendHeader("Access-Control-Allow-Origin", "*");
-        server.send(200, "text/plain", "{\"status\":\"ok\"}");
-    }
-}
-
-void handleNotFound()
-{
-    digitalWrite(led, 1);
-    String message = "File Not Found\n\n";
-    message += "URI: ";
-    message += server.uri();
-    message += "\nMethod: ";
-    message += (server.method() == HTTP_GET) ? "GET" : "POST";
-    message += "\nArguments: ";
-    message += server.args();
-    message += "\n";
-
-    for (uint8_t i = 0; i < server.args(); i++)
-    {
-        message += " " + server.argName(i) + ": " + server.arg(i) + "\n";
-    }
-    server.sendHeader("Access-Control-Allow-Origin", "*");
-    server.send(404, "text/plain", message);
-    digitalWrite(led, 0);
-}
-
-void handleConfigGains()
-{
-    if (server.method() != HTTP_POST)
-    {
-        digitalWrite(led, 1);
-        server.send(405, "text/plain", "Method Not Allowed");
-        digitalWrite(led, 0);
-    }
-    else
-    {
-
-        for (uint8_t i = 0; i < server.args(); i++)
-        {
-            deserializeJson(doc, server.arg(i));
-            for (uint8_t j = 0; j < 6; j++)
-            {
-                LoadAnalogInputGains(server.argName(i).toInt(), j, doc[String(j)]);
-            }
-        }
-    }
-    server.sendHeader("Access-Control-Allow-Origin", "*");
-    server.send(200, "text/plain", "Listo ;) ");
-}
-
-void handleReadGains()
-{
-    server.sendHeader("Access-Control-Allow-Origin", "*");
-    server.send(200, "text/plain", ReturnAnalogInputGains());
-}
-
-void handleStatus()
-{
-    String status = "[{\"data\":";
-    status += printStatusADS();
-    status += ", " + printStatusExpander();
-    status += "]}";
-    server.sendHeader("Access-Control-Allow-Origin", "*");
-    server.send(200, "text/plain", status);
-}
+#include "Server_Handlers.h"
 
 void setGetsPosts()
 {
@@ -226,7 +80,7 @@ void setup()
     startAnalogInputs();
     startExpanders();
     startVoltageOutputs();
-    connectToEthernet();
+    // connectToEthernet();
 }
 
 void loop()
